+++ pywwt/__init__.py
@@ -1,6 +1,7 @@
 from ._version import version_info, __version__  # noqa
 from .core import *  # noqa
 from .annotation import *  # noqa
+from .jupyter_server import load_jupyter_server_extension  # noqa
 
 
 def _jupyter_nbextension_paths():
@@ -8,3 +9,7 @@ def _jupyter_nbextension_paths():
              'src': 'static',
              'dest': 'pywwt',
              'require': 'pywwt/extension'}]
+
+
+def _jupyter_server_extension_paths():
+    return [{"module": "pywwt"}]